🧼 Warnings as Errors

+++ Modules/Dementia/Source/Math/Quaternion.cpp
@@ -21,19 +21,19 @@ Vector3 Quaternion::ToEulerAngles(const Quaternion& InQuat)
 	// roll (x-axis rotation)
 	double sinr_cosp = 2 * (q.w * q.x + q.y * q.z);
 	double cosr_cosp = 1 - 2 * (q.x * q.x + q.y * q.y);
-	angles.x = std::atan2(sinr_cosp, cosr_cosp);
+	angles.x = static_cast<float>(std::atan2(sinr_cosp, cosr_cosp));
 
 	// pitch (y-axis rotation)
 	double sinp = 2 * (q.w * q.y - q.z * q.x);
 	if (std::abs(sinp) >= 1)
-		angles.y = std::copysign(M_PI / 2, sinp); // use 90 degrees if out of range
+		angles.y = static_cast<float>(std::copysign(M_PI / 2, sinp)); // use 90 degrees if out of range
 	else
-		angles.y = std::asin(sinp);
+		angles.y = static_cast<float>(std::asin(sinp));
 
 	// yaw (z-axis rotation)
 	double siny_cosp = 2 * (q.w * q.z + q.x * q.y);
 	double cosy_cosp = 1 - 2 * (q.y * q.y + q.z * q.z);
-	angles.z = std::atan2(siny_cosp, cosy_cosp);
+	angles.z = static_cast<float>(std::atan2(siny_cosp, cosy_cosp));
 
 	return angles;
 }
