+++ Source/UI/UIWindow.cpp
@@ -34,12 +34,12 @@ ultralight::WindowListener* UIWindow::listener()
 
 uint32_t UIWindow::width() const
 {
-	return m_window->GetSize().x;
+	return static_cast<uint32_t>(m_window->GetSize().x);
 }
 
 uint32_t UIWindow::height() const
 {
-	return m_window->GetSize().y;
+	return static_cast<uint32_t>(m_window->GetSize().y);
 }
 
 bool UIWindow::is_fullscreen() const
